Move the new tree_centroid to networkx/algorithms/tree/
