+++ doc/reference/algorithms/tree.rst
@@ -38,6 +38,8 @@ Distance Measures
    :toctree: generated/
 
    center
+   centroid
+
 
 Encoding and decoding
 ---------------------
